requirements.txt
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
django-tailwind==4.2.0
mysqlclient==2.2.1
pillow==10.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
